code for sliding window is modularizedand appears to work.
The next thing to work on would be to make the dynamcics
for q-dot, p-dot, and u-dot as inputs. (i.e. make the function
names inputs.

In the code in Control_Law_Propagator.ipynb, that would mean making
H_T_p and H_T_q and Q_u as inputs to the function "sliding_window".

After that write unit tests.