Through some testing, I discovered that the qp_rhs_H_mf should actually take in ALL of the states, not just the ones that pertain to the agent.  So, this means that the dimensions of the inputs should be equal to the total number of states in the entire system, rather than equal to just the dimension of the number of states pertaining to that agent.
Outside of each qp_rhs_H_mf method, (but inside of the qp_rhs method)
we will construct the vector combining quenched and non-quenched states
to make a vector with dimension equal to the number of states in the
entire system.  Then, this vector will be passed into all of the subsequent
H_mf methods.