Added notebook Mean_Field_Library_tests.ipynb in order to run the same thing that Mean_Field.ipynb did, except run it using the methods in the python files instead of in a notebook.
NOw, start deleting the methods in the notebook in order to truly
test the python files.

also made some minor clean-ups in the python files:
    - backboard.py
    - synchronizer.py
    - numerical_propagator.py

The only things that should live in the notebook at the end of this
are the things that will go in the config file (e.g. the definition
of class Agent, etc.)