After reminding myself that u_s is not needed for H_l_u, nor for H_l_nou, I removed it from the input parameters for those methods in batteryAgent.py.
NOTE: This is not true for H_mf_u and H_mf_nou - those methods still need
u_mf because the values in u_mf are quenched and hence can be used in the
terms in those expressions.

Also keep in mind that lambda_l is needed as an input to H_l_nou, even
though currently we are not using that, but it is not needed as an
input into H_l_u.